requirements.txt
Appium-Python-Client
requests==2.*
